// src/database/entities/enums.ts
export enum UserRole {
  USER = "user",
  ADMIN = "admin",
}

export enum ApiKeyType {
  LIVE = "live",
  TEST = "test",
}

export enum PlanInterval {
  DAILY = "daily",
  MONTHLY = "monthly",
  YEARLY = "yearly",
}

export enum UsageStatus {
  SUCCESS = "success",
  FAILED = "failed",
  PENDING = "pending",
}